Refactor custom sync
