+++ android/quest/src/main/res/values/strings.xml
@@ -144,5 +144,4 @@
     <string name="error_rendering_profile">Error rendering profile</string>
     <string name="questionnaire_submission_confirmation_message">Are you sure you want to submit?</string>
     <string name="questionnaire_submission_confirmation_title">You are about to submit</string>
-    <string name="syncing_custom_resources_toast">Syncing custom Resources...</string>
 </resources>
